extract out ActionNav
